fixes navigation bug between stakes and gauges
